build for gcc3.2

+++ panda/src/chan/animControl.cxx
@@ -258,7 +258,7 @@ pose(int frame) {
   nassertv(num_frames > 0);
 
   // Modulo the number of frames.
-  frame = frame - cfloor(frame / num_frames) * num_frames;
+  frame = (int)(frame - cfloor(frame / num_frames) * num_frames);
   nassertv(frame >= 0 && frame < num_frames);
   _as_of_time = ClockObject::get_global_clock()->get_frame_time();
   _playing = false;
